test_web/requirements.txt
pytest==8.0.0
playwright==1.40.0
python-dotenv==1.0.0
